% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{.randomColors}
\alias{.randomColors}
\title{Random colors}
\usage{
.randomColors(ncolors)
}
\arguments{
\item{ncolors}{Integer. Number of colors to generate randomly.}
}
\value{
Random colors.
}
\description{
The function returns a group of random colors.
}
\keyword{internal}
