% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{DownOverSampleEvenlyBatches}
\alias{DownOverSampleEvenlyBatches}
\title{Down- and oversample data evenly batches}
\usage{
DownOverSampleEvenlyBatches(x, batch, n = 50)
}
\arguments{
\item{x}{A character or numeric vector of data to down-and oversample.}

\item{batch}{A character vector with batch labels corresponding to \code{x}.}

\item{n}{How many cells to include per cluster.}
}
\value{
a list containing the output of the LiblineaR prediction
}
\description{
The function down- and over-samples cluster cells evenly by batch.¸
}
\keyword{downsampling}
\keyword{oversampling}
