% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{FindBatchKNN}
\alias{FindBatchKNN}
\title{Find batch k nearest neighbors}
\usage{
FindBatchKNN(idx, group, prob, k = 10)
}
\arguments{
\item{idx}{A numeric vector with the cell ids to retrieve from the data set.}

\item{group}{A character vector with batch labels corresponding to \code{idx}.}

\item{prob}{A numeric vector with cell probabilities corresponding to \code{idx}.}

\item{k}{The number of nearest neighbors to search for. Default is \code{10}.}
}
\value{
a list containing the k nearest neighbors for every cell queried
}
\description{
The function finds batch k nearest neighbors for the cell with the highest
probability for every batch.
}
\keyword{internal}
