% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrationMethods.R
\name{IntegrateDimRed}
\alias{IntegrateDimRed}
\alias{IntegrateDimRed.SingleCellExperiment}
\alias{IntegrateDimRed,SingleCellExperiment-method}
\title{Integrate low dimensional reduction}
\usage{
IntegrateDimRed.SingleCellExperiment(object, pc, ld, cluster.method)

\S4method{IntegrateDimRed}{SingleCellExperiment}(object, pc = 50, ld = 9, cluster.method = "NNgraph")
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class.}

\item{pc}{Number of principal components (integer) to compute. By default 
\code{50}.}

\item{ld}{Number of linear discriminants to compute (integer). By default 
\code{9}.}

\item{cluster.method}{Cluster method to cluster PCA in order to give the 
clustering result for supervised Linear Discriminant Analysis (character). 
One of \code{"kmeans++"} (kmeans optimization implement in `flexclust`) or 
\code{"NNgraph"} (graph based clustering implemented in `scran`). By default 
\code{"NNgraph"}.}
}
\value{
A clustering result where every cluster given was split randomly 
into \code{k} clusters.
}
\description{
Integrates low dimensional reduction, i.e., PCA, through supervised LDA.
}
\keyword{LDA}
\keyword{PCA}
\keyword{dimensional}
\keyword{reduction}
