% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\docType{data}
\name{pbmc_10Xassays}
\alias{pbmc_10Xassays}
\title{A toy SCE dataset of two 10X PBMCs 3' assays downsampled to 2K cells.}
\format{
SCE object
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/pbmc6k}

\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/2.1.0/pbmc8k}
}
\usage{
data(pbmc_10Xassays)
pbmc_10Xassays
}
\description{
Two 10X PBMCs 3' assays, V1 and V2, downsampled to 2K cells (1K per assay).  
Both datasets have been downloaded from 10X website (see source below). The 
PBMC dataset V1 corresponds to sample pbmc6k and V2 pbmc8k. Cells were  
annotated using the annotations provided by Korsunsky et al., 2019 (PMID:31740819; 
Source Data Figure 4 file).
}
\examples{
data(pbmc_10Xassays)
pbmc_10Xassays
}
\keyword{datasets}
