% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{PrepareILoReg2}
\alias{PrepareILoReg2}
\alias{PrepareILoReg2.SingleCellExperiment}
\alias{PrepareILoReg2,SingleCellExperiment-method}
\title{Prepare \code{SingleCellExperiment} object for \code{ILoReg2} analysis}
\usage{
PrepareILoReg2.SingleCellExperiment(object)

\S4method{PrepareILoReg2}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}
}
\value{
an object of \code{SingleCellExperiment} class
}
\description{
This function prepares the \code{SingleCellExperiment} object for
\code{ILoReg2} analysis. The only required input is an object of class
\code{SingleCellExperiment} with at least data in the \code{logcounts} slot.
}
\examples{
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays = list(logcounts = pbmc3k_500))
sce <- PrepareILoReg2(sce)

}
\keyword{clean}
\keyword{data}
\keyword{iloreg}
\keyword{normalized}
\keyword{prepare}
