% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrationMethods.R
\name{SamplePCABatchCells}
\alias{SamplePCABatchCells}
\title{Sample cells based on principal components batch distribution}
\usage{
SamplePCABatchCells(
  data,
  batch,
  q.split = 0.5,
  p = 30,
  use.pc = "PC1",
  center = TRUE,
  scale. = TRUE
)
}
\arguments{
\item{data}{Data to compute PCA and sample cells from. Rows and columns 
should represent cells and features/genes, respectively.}

\item{batch}{Batch cell label identity (character) matching cells giving in 
\code{data}.}

\item{q.split}{Split (cell) batch principal component distribution by this 
quantile (numeric). By default {0.5}, i.e., median.}

\item{p}{Number of principal components to compute (integer). By default 
\code{30}.}

\item{use.pc}{Which principal component should be used for sampling cells per
batch. By default \code{"PC1"}, i.e., first principal component is used.}

\item{center}{Should the features/genes given in \code{data} centered before
performing the PCA (logical). By default \code{TRUE}.}

\item{scale.}{Should the features/genes given in \code{data} scaled before
performing the PCA (logical). By default \code{TRUE}.}
}
\value{
A factor with cell cluster identities (two clusters).
}
\description{
Samples cells based on batch distributions along one principal component
}
\keyword{PCA}
\keyword{batch}
\keyword{sample}
