% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{ScaleByBatch}
\alias{ScaleByBatch}
\title{Scale sparse matrix by features (column) by batch}
\usage{
ScaleByBatch(x, batch)
}
\arguments{
\item{x}{A matrix of class `dgCMatrix`. Cells by features (rows x columns).}

\item{batch}{A character vector with batch labels corresponding to the cells
given in \code{x}. The character batch labels need to be named
with the cells names given in the rows of \code{x}.}
}
\value{
A scaled matrix of class `dgCMatrix`.
}
\description{
Scales features by batch
}
\keyword{internal}
