% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrationMethods.R
\name{AggregateDataByBatch}
\alias{AggregateDataByBatch}
\alias{AggregateDataByBatch.SingleCellExperiment}
\alias{AggregateDataByBatch,SingleCellExperiment-method}
\title{Aggregates gene expression by cell clusters, per batch if provided.}
\usage{
AggregateDataByBatch.SingleCellExperiment(object, batch.label, nhvg, p, ...)

\S4method{AggregateDataByBatch}{SingleCellExperiment}(object, batch.label, nhvg = 2000L, p = 30L, ...)
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class.}

\item{batch.label}{Cluster identities vector corresponding to the cells in 
\code{mtx}.}

\item{nhvg}{Integer of the number of highly variable genes to select. By default 
\code{2000}.}

\item{p}{Integer. By default \code{30}.}

\item{...}{Parameters to be passed to \code{ClusterCells()} function.}
}
\value{
A SingleCellExperiment object with gene expression aggregated by clusters.
}
\description{
The function aggregates gene expression by cell clusters, per batch if provided.
}
\keyword{aggregated}
\keyword{batches}
\keyword{expression}
\keyword{gene}
