% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellStateMethods.R
\name{CellBinsFeatureCorrelation}
\alias{CellBinsFeatureCorrelation}
\alias{CellBinsFeatureCorrelation.SingleCellExperiment}
\alias{CellBinsFeatureCorrelation,SingleCellExperiment-method}
\title{Cell bins feature correlation}
\usage{
CellBinsFeatureCorrelation.SingleCellExperiment(object, labels, method)

\S4method{CellBinsFeatureCorrelation}{SingleCellExperiment}(object, labels = NULL, method = "pearson")
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class obtained with the 
function \code{BinCellClusterProbability()}.}

\item{labels}{Character of label(s) from the label provided to the function 
\code{BinCellClusterProbability()}. By default \code{NULL}, i.e., all labels
are used.}

\item{method}{Character specifying the correlation method to use. One of 
\code{"pearson"}, \code{"kendall"} or \code{"spearman"}. By default \code{"pearson"}
is used.}
}
\value{
A data frame with the correlation coefficient for each feature (rows) 
across labels (columns).
}
\description{
Correlation between cell bins for the given labels and features.
}
\keyword{Cell}
\keyword{bins}
\keyword{correlation}
\keyword{feature}
