% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellStateMethods.R
\name{CellClusterProbabilityDistribution}
\alias{CellClusterProbabilityDistribution}
\alias{CellClusterProbabilityDistribution.SingleCellExperiment}
\alias{CellClusterProbabilityDistribution,SingleCellExperiment-method}
\title{Cell cluster probability distribution}
\usage{
CellClusterProbabilityDistribution.SingleCellExperiment(
  object,
  label,
  group,
  probability
)

\S4method{CellClusterProbabilityDistribution}{SingleCellExperiment}(
  object,
  label,
  group,
  probability = "scaled_mean_probs"
)
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class with aggregated 
cell cluster probability available in \code{colData(object)}, which can be 
obtained after running \code{SummariseCellClusterProbability()}.}

\item{label}{Character specifying the \code{colData} variable to use as cell 
type/cluster label.}

\item{group}{Character specifying the \code{colData} variable to use as 
categorical group variable.}

\item{probability}{Character specifying the aggregated cell cluster probability 
variable available in \code{colData}, used to plot its distribution. One of 
\code{"mean_probs"}, \code{"scaled_mean_probs"}, \code{"median_probs"}, 
\code{"scaled_median_probs"}. The availability of these variables in \code{colData} 
depends on the parameters given to the function \code{SummariseCellClusterProbability()} 
beforehand. By default assumes that \code{"scaled_mean_probs"} is available in
\code{colData}, which is only true if \code{SummariseCellClusterProbability()} 
function was run with \code{funs = "mean"} and \code{scale.funs = TRUE}.}
}
\value{
A plot of class \code{ggplot}.
}
\description{
Plot cell cluster probability distribution per label by group.
}
\examples{
# Packages
library("SingleCellExperiment")

# Prepare data
pbmc_10Xassays <- PrepareData(object = pbmc_10Xassays)

# Multi-level integration - 'L = 4' just for highlighting purposes; use 'L=50' or greater
set.seed(123)
pbmc_10Xassays <- RunParallelDivisiveICP(object = pbmc_10Xassays, batch.label = "batch", L = 4, threads = 1)

# Summarise cell cluster probability
pbmc_10Xassays <- SummariseCellClusterProbability(object = pbmc_10Xassays, icp.round = 4) # save result in 'colData'

# Search for differences in probabilities across group(s) 
# give an interesting variable to the "group" parameter instead "batch" - given for illustrative purposes only
prob.dist <- CellClusterProbabilityDistribution(object = pbmc_10Xassays, label = "cell_type", group = "batch", probability = "scaled_mean_probs")
prob.dist # print plot

}
\keyword{Distribution}
\keyword{cell}
\keyword{cluster}
\keyword{probability}
