% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellStateMethods.R
\name{CellClusterProbabilityDistribution}
\alias{CellClusterProbabilityDistribution}
\alias{CellClusterProbabilityDistribution.SingleCellExperiment}
\alias{CellClusterProbabilityDistribution,SingleCellExperiment-method}
\title{Cell cluster probability distribution}
\usage{
CellClusterProbabilityDistribution.SingleCellExperiment(
  object,
  label,
  group,
  probability
)

\S4method{CellClusterProbabilityDistribution}{SingleCellExperiment}(
  object,
  label,
  group,
  probability = "scaled_mean_probs"
)
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class with aggregated 
cell cluster probability available in \code{colData(object)}, which can be 
obtained after running \code{SummariseCellClusterProbability()}.}

\item{label}{Character specifying the \code{colData} variable to use as cell 
type/cluster label.}

\item{group}{Character specifying the \code{colData} variable to use as 
categorical group variable.}

\item{probability}{Character specifying the aggregated cell cluster probability 
variable available in \code{colData}, used to plot its distribution. One of 
\code{"mean_probs"}, \code{"scaled_mean_probs"}, \code{"median_probs"}, 
\code{"scaled_median_probs"}. The availability of these variables in \code{colData} 
depends on the parameters given to the function \code{SummariseCellClusterProbability()} 
beforehand. By default assumes that \code{"scaled_mean_probs"} is available in
\code{colData}, which is only true if \code{SummariseCellClusterProbability()} 
function was run with \code{funs = "mean"} and \code{scale.funs = TRUE}.}
}
\value{
A plot of class \code{ggplot}.
}
\description{
Plot cell cluster probability distribution per label by group
}
\keyword{Distribution}
\keyword{cell}
\keyword{cluster}
\keyword{probability}
