% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{ClusterCells}
\alias{ClusterCells}
\title{Cluster cells}
\usage{
ClusterCells(object, nclusters = 150, use.emb = TRUE, emb.name = "PCA")
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class.}

\item{nclusters}{Cluster the cells into n clusters. Ignored if the number of 
cells in \code{object} is lower or equal to \code{nclusters}.}

\item{use.emb}{Should the embedding be used to cluster or the log-transformed 
data. By default \code{TRUE}.}

\item{emb.name}{Which embedding to use. By default \code{"PCA"}.}
}
\value{
Clusters.
}
\description{
The function clusters cells with the K-means++ algorithm
}
\keyword{cluster}
