% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{FindClusterBatchKNN}
\alias{FindClusterBatchKNN}
\title{Find batch k nearest neighbors per cluster}
\usage{
FindClusterBatchKNN(preds, probs, batch, k = 10)
}
\arguments{
\item{preds}{A numeric vector with the cell cluster predictions.}

\item{probs}{A numeric matrix with cell cluster probabilities.}

\item{batch}{A numeric vector with cell probabilities corresponding to \code{idx}.}

\item{k}{The number of nearest neighbors to search for. Default is \code{10}.}
}
\value{
a list containing the k nearest neighbors for every cluster
}
\description{
The function finds batch k nearest neighbors for the cell with the highest 
probability for every batch per cluster.
}
\keyword{knn}
