% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{GetCellClusterProbability}
\alias{GetCellClusterProbability}
\alias{GetCellClusterProbability.SingleCellExperiment}
\alias{GetCellClusterProbability,SingleCellExperiment-method}
\title{Get ICP cell cluster probability}
\usage{
GetCellClusterProbability.SingleCellExperiment(
  object,
  icp.run,
  icp.round,
  concatenate
)

\S4method{GetCellClusterProbability}{SingleCellExperiment}(
  object,
  icp.run = NULL,
  icp.round = NULL,
  concatenate = TRUE
)
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class with ICP cell 
cluster probability tables saved in \code{metadata(object)$iloreg$joint.probability}. 
After running one of \code{RunParallelICP} or \code{RunParallelDivisiveICP}.}

\item{icp.run}{ICP run(s) to retrieve from \code{metadata(object)$iloreg$joint.probability}. 
By default \code{NULL}, i.e., all are retrieved. Specify a numeric vector to 
retrieve a specific set of tables.}

\item{icp.round}{ICP round(s) to retrieve from \code{metadata(object)$iloreg$joint.probability}. 
By default \code{NULL}, i.e., all are retrieved. Only relevant if probabilities
were obtained with the function \code{RunParallelDivisiveICP}, i.e., divisive ICP
was performed. Otherwise it is ignored and internally assumed as \code{icp.round = 1}, 
i.e., only one round.}

\item{concatenate}{Concatenate list of ICP cell cluster probability tables retrieved. 
By default \code{TRUE}, i.e., the list of ICP cell cluster probability tables is
concatenated.}
}
\value{
A list with ICP cell cluster probability tables or a matrix with 
concatenated tables.
}
\description{
Get ICP cell cluster probability table(s)
}
\keyword{Cell}
\keyword{cluster}
\keyword{probability}
