% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{GetFeatureCoefficients}
\alias{GetFeatureCoefficients}
\alias{GetFeatureCoefficients.SingleCellExperiment}
\alias{GetFeatureCoefficients,SingleCellExperiment-method}
\title{Get feature coefficients}
\usage{
GetFeatureCoefficients.SingleCellExperiment(
  object,
  icp.run = NULL,
  icp.round = NULL
)

\S4method{GetFeatureCoefficients}{SingleCellExperiment}(object, icp.run = NULL, icp.round = NULL)
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class with ICP cell 
cluster probability tables saved in \code{metadata(object)$iloreg$joint.probability}. 
After running one of \code{RunParallelICP} or \code{RunParallelDivisiveICP}.}

\item{icp.run}{ICP run(s) to retrieve from \code{metadata(object)$iloreg$joint.probability}. 
By default \code{NULL}, i.e., all are retrieved. Specify a numeric vector to 
retrieve a specific set of tables.}

\item{icp.round}{ICP round(s) to retrieve from \code{metadata(object)$iloreg$joint.probability}. 
By default \code{NULL}, i.e., all are retrieved. Only relevant if probabilities
were obtained with the function \code{RunParallelDivisiveICP}, i.e., divisive ICP
was performed. Otherwise it is ignored and internally assumed as \code{icp.round = 1}, 
i.e., only one round.}
}
\value{
A list of feature coefficient weights per cluster per ICP run/round.
}
\description{
Get feature coefficients from ICP models.
}
\keyword{Feature}
\keyword{coefficients}
\keyword{weights}
