% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{GetFeatureCoefficients}
\alias{GetFeatureCoefficients}
\alias{GetFeatureCoefficients.SingleCellExperiment}
\alias{GetFeatureCoefficients,SingleCellExperiment-method}
\title{Get feature coefficients}
\usage{
GetFeatureCoefficients.SingleCellExperiment(
  object,
  icp.run = NULL,
  icp.round = NULL
)

\S4method{GetFeatureCoefficients}{SingleCellExperiment}(object, icp.run = NULL, icp.round = NULL)
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class with ICP cell 
cluster probability tables saved in \code{metadata(object)$coralysis$joint.probability}. 
After running \code{RunParallelDivisiveICP}.}

\item{icp.run}{ICP run(s) to retrieve from \code{metadata(object)$coralysis$joint.probability}. 
By default \code{NULL}, i.e., all are retrieved. Specify a numeric vector to 
retrieve a specific set of tables.}

\item{icp.round}{ICP round(s) to retrieve from \code{metadata(object)$coralysis$joint.probability}. 
By default \code{NULL}, i.e., all are retrieved.}
}
\value{
A list of feature coefficient weights per cluster per ICP run/round.
}
\description{
Get feature coefficients from ICP models.
}
\examples{
# Import package
library("SingleCellExperiment")

# Prepare SCE object for analysis
pbmc_10Xassays <- PrepareData(pbmc_10Xassays)

# Multi-level integration - 'L = 4' just for highlighting purposes; use 'L=50' or greater
set.seed(123)
pbmc_10Xassays <- RunParallelDivisiveICP(object = pbmc_10Xassays, batch.label = "batch", L = 4, threads = 1) 

# Run PCA 
pbmc_10Xassays <- RunPCA(pbmc_10Xassays, p = 10)

# GetFeatureCoefficients
gene_coefficients_icp_2_4 <- GetFeatureCoefficients(object = pbmc_10Xassays, icp.run = 2, icp.round = 4)
head(gene_coefficients_icp_2_4$icp_8)

}
\keyword{Feature}
\keyword{coefficients}
\keyword{weights}
