% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{GetFeatureCoefficients}
\alias{GetFeatureCoefficients}
\alias{GetFeatureCoefficients.SingleCellExperiment}
\alias{GetFeatureCoefficients,SingleCellExperiment-method}
\title{Get feature coefficients}
\usage{
GetFeatureCoefficients.SingleCellExperiment(
  object,
  icp.run = NULL,
  icp.round = NULL
)

\S4method{GetFeatureCoefficients}{SingleCellExperiment}(object, icp.run = NULL, icp.round = NULL)
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class with ICP cell 
cluster probability tables saved in \code{metadata(object)$coralysis$joint.probability}. 
After running \code{RunParallelDivisiveICP}.}

\item{icp.run}{ICP run(s) to retrieve from \code{metadata(object)$coralysis$joint.probability}. 
By default \code{NULL}, i.e., all are retrieved. Specify a numeric vector to 
retrieve a specific set of tables.}

\item{icp.round}{ICP round(s) to retrieve from \code{metadata(object)$coralysis$joint.probability}. 
By default \code{NULL}, i.e., all are retrieved.}
}
\value{
A list of feature coefficient weights per cluster per ICP run/round.
}
\description{
Get feature coefficients from ICP models.
}
\examples{
\dontrun{
# Import package
suppressPackageStartupMessages(library("SingleCellExperiment"))

# Import data from Zenodo
data.url <- "https://zenodo.org/records/14845751/files/pbmc_10Xassays.rds?download=1"
sce <- readRDS(file = url(data.url))

# Multi-level integration (just for highlighting purposes; use default parameters)
set.seed(123)
sce <- RunParallelDivisiveICP(object = sce, batch.label = "batch", 
                              k = 4, L = 10, C = 1, d = 0.5, 
                              train.with.bnn = FALSE, use.cluster.seed = FALSE,
                              build.train.set = FALSE, ari.cutoff = 0.1, 
                              threads = 2)

# GetFeatureCoefficients
gene_coefficients_icp_2_4 <- GetFeatureCoefficients(object = sce, 
                                                    icp.run = 2, icp.round = 4)
head(gene_coefficients_icp_2_4$icp_8)
}

}
\keyword{Feature}
\keyword{coefficients}
\keyword{weights}
