% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizMethods.R
\name{HeatmapFeatures}
\alias{HeatmapFeatures}
\alias{HeatmapFeatures.SingleCellExperiment}
\alias{HeatmapFeatures,SingleCellExperiment-method}
\title{Heatmap visualization of the expression of features by clusters}
\usage{
HeatmapFeatures.SingleCellExperiment(
  object,
  clustering.label,
  features,
  use.color,
  seed.color,
  ...
)

\S4method{HeatmapFeatures}{SingleCellExperiment}(
  object,
  clustering.label,
  features,
  use.color = NULL,
  seed.color = 123,
  ...
)
}
\arguments{
\item{object}{of \code{SingleCellExperiment} class}

\item{clustering.label}{A variable name (of class \code{character}) available 
in the cell metadata \code{colData(object)} with the clustering labels 
(\code{character} or \code{factor}) to use.}

\item{features}{Feature names to plot by cluster (\code{character}) matching 
\code{row.names(object)}.}

\item{use.color}{Character specifying the colors for the clusters. By default 
\code{NULL}, i.e., colors are randomly chosen based on the seed given at 
\code{seed.color}. It is overwritten in case the argument \code{annotation_colors} 
is provided.}

\item{seed.color}{Seed to randomly select colors for the clusters. By default 
\code{123}. It is overwritten in case the argument \code{annotation_colors} 
is provided.}

\item{...}{Parameters to pass to \code{pheatmap::pheatmap} function.}
}
\value{
nothing
}
\description{
The \code{HeatmapFeatures} function draws a heatmap of features 
by cluster identity.
}
\examples{
# Import package
library("SingleCellExperiment")

# Prepare SCE object for analysis
pbmc_10Xassays <- PrepareData(pbmc_10Xassays)

# Multi-level integration - 'L = 4' just for highlighting purposes; use 'L=50' or greater
set.seed(123)
pbmc_10Xassays <- RunParallelDivisiveICP(object = pbmc_10Xassays, batch.label = "batch", L = 4, threads = 1) 

# Run PCA 
set.seed(125) # to ensure reproducibility for the default 'irlba' method
pbmc_10Xassays <- RunPCA(pbmc_10Xassays, p = 10)

# Plot features by cell type (give a clustering instead)
genes <- c("CD8A", "CCR7", "GZMA", "MS4A1", "CD27")
HeatmapFeatures(object = pbmc_10Xassays, clustering.label = "cell_type", features = genes) # log-normalized
HeatmapFeatures(object = pbmc_10Xassays, clustering.label = "cell_type", features = genes, scale = "row") # scale genes

}
\keyword{feature}
\keyword{grouped}
\keyword{heatmap}
