% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{MajorityVotingFeatures}
\alias{MajorityVotingFeatures}
\alias{MajorityVotingFeatures.SingleCellExperiment}
\alias{MajorityVotingFeatures,SingleCellExperiment-method}
\title{Majority voting features by label}
\usage{
MajorityVotingFeatures.SingleCellExperiment(object, label)

\S4method{MajorityVotingFeatures}{SingleCellExperiment}(object, label)
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class with ICP cell 
cluster probability tables saved in \code{metadata(object)$iloreg$joint.probability}. 
After running one of \code{RunParallelICP} or \code{RunParallelDivisiveICP}.}

\item{label}{Label of interest available in \code{colData(object)}.}
}
\value{
A list of with a list of data frames with feature weights per label and a data frame with a summary by label.
}
\description{
Get ICP feature coefficients for a label of interest by majority voting label across ICP clusters.
}
\keyword{Majority}
\keyword{coefficients}
\keyword{feature}
\keyword{voting}
\keyword{weights}
