% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{PCAElbowPlot}
\alias{PCAElbowPlot}
\alias{PCAElbowPlot.SingleCellExperiment}
\alias{PCAElbowPlot,SingleCellExperiment-method}
\title{Elbow plot of the standard deviations of the principal components}
\usage{
PCAElbowPlot.SingleCellExperiment(object, dimred.name, return.plot)

\S4method{PCAElbowPlot}{SingleCellExperiment}(object, dimred.name = "PCA", return.plot = FALSE)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object obtained after running 
\code{RunParallelDivisiveICP}.}

\item{dimred.name}{Dimensional reduction name of the PCA to select from 
\code{reducedDimNames(object)}. By default \code{"PCA"}.}

\item{return.plot}{logical indicating if the ggplot2 object should be returned.
By default \code{FALSE}.}
}
\value{
A ggplot2 object, if \code{return.plot=TRUE}.
}
\description{
Draw an elbow plot of the standard deviations of the principal 
components to deduce an appropriate value for \code{p}.
}
\examples{
# Import package
library("SingleCellExperiment")

# Prepare SCE object for analysis
pbmc_10Xassays <- PrepareData(pbmc_10Xassays)

# Multi-level integration - 'L = 4' just for highlighting purposes; use 'L=50' or greater
set.seed(123)
pbmc_10Xassays <- RunParallelDivisiveICP(object = pbmc_10Xassays, batch.label = "batch", L = 4, threads = 1) 

# Run PCA 
set.seed(125) # to ensure reproducibility for the default 'irlba' method
pbmc_10Xassays <- RunPCA(pbmc_10Xassays, p = 10)

# Plot Elbow
PCAElbowPlot(pbmc_10Xassays)

}
\keyword{PCA}
\keyword{elbow}
\keyword{plot}
