% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizMethods.R
\name{PlotExpression}
\alias{PlotExpression}
\alias{PlotExpression.SingleCellExperiment}
\alias{PlotExpression,SingleCellExperiment-method}
\title{Plot dimensional reduction feature expression}
\usage{
PlotExpression.SingleCellExperiment(
  object,
  color.by,
  dimred,
  scale.values,
  color.scale,
  plot.theme,
  legend.title,
  point.size,
  point.stroke
)

\S4method{PlotExpression}{SingleCellExperiment}(
  object,
  color.by,
  dimred = tail(reducedDimNames(object), n = 1),
  scale.values = FALSE,
  color.scale = "inferno",
  plot.theme = theme_classic(),
  legend.title = color.by,
  point.size = 1,
  point.stroke = 1
)
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class.}

\item{color.by}{Categorical variable available in \code{colData(object)} to 
plot.}

\item{dimred}{Dimensional reduction available in \code{ReducedDimNames(object)}
to plot. By default the last dimensional reduction in the object is used.}

\item{color.scale}{Character of color scale palette to be passed to 
\code{ggplot2::scale_color_viridis_c}. By default \code{inferno}. Other palettes
are also available such as \code{viridis}.}

\item{plot.theme}{Plot theme available in \code{ggplot2}. By default \code{theme_classic()}.}

\item{legend.title}{Legend title. By default the same as given at \code{color.by}.}

\item{point.size}{Size of points. By default \code{1}.}

\item{point.stroke}{Size of stroke. By default \code{1}.}
}
\value{
A plot of class \code{ggplot}.
}
\description{
Plot feature expression in dimensional reduction.
}
\keyword{Dimensional}
\keyword{reduction}
\keyword{visualization}
