% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizMethods.R
\name{PlotExpression}
\alias{PlotExpression}
\alias{PlotExpression.SingleCellExperiment}
\alias{PlotExpression,SingleCellExperiment-method}
\title{Plot dimensional reduction feature expression}
\usage{
PlotExpression.SingleCellExperiment(
  object,
  color.by,
  dimred,
  scale.values,
  color.scale,
  plot.theme,
  legend.title,
  point.size,
  point.stroke
)

\S4method{PlotExpression}{SingleCellExperiment}(
  object,
  color.by,
  dimred = tail(reducedDimNames(object), n = 1),
  scale.values = FALSE,
  color.scale = "inferno",
  plot.theme = theme_classic(),
  legend.title = color.by,
  point.size = 1,
  point.stroke = 1
)
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class.}

\item{color.by}{Categorical variable available in \code{colData(object)} to 
plot.}

\item{dimred}{Dimensional reduction available in \code{ReducedDimNames(object)}
to plot. By default the last dimensional reduction in the object is used.}

\item{color.scale}{Character of color scale palette to be passed to 
\code{ggplot2::scale_color_viridis_c}. By default \code{inferno}. Other palettes
are also available such as \code{viridis}.}

\item{plot.theme}{Plot theme available in \code{ggplot2}. By default \code{theme_classic()}.}

\item{legend.title}{Legend title. By default the same as given at \code{color.by}.}

\item{point.size}{Size of points. By default \code{1}.}

\item{point.stroke}{Size of stroke. By default \code{1}.}
}
\value{
A plot of class \code{ggplot}.
}
\description{
Plot feature expression in dimensional reduction.
}
\examples{
# Import package
library("SingleCellExperiment")

# Prepare SCE object for analysis
pbmc_10Xassays <- PrepareData(pbmc_10Xassays)

# Multi-level integration - 'L = 4' just for highlighting purposes; use 'L=50' or greater
set.seed(123)
pbmc_10Xassays <- RunParallelDivisiveICP(object = pbmc_10Xassays, batch.label = "batch", L = 4, threads = 1) 

# Run PCA 
set.seed(125) # to ensure reproducibility for the default 'irlba' method
pbmc_10Xassays <- RunPCA(pbmc_10Xassays, p = 10)

# Run UMAP
set.seed(123)
pbmc_10Xassays <- RunUMAP(pbmc_10Xassays, dimred.type = "PCA")

# Plot cell cluster probabilities - mean
# possible options: "mean_probs", "median_probs", "scaled_median_probs" 
pbmc_10Xassays <- SummariseCellClusterProbability(object = pbmc_10Xassays, icp.round = 4, scale.funs = TRUE, save.in.sce = TRUE) # save result in 'colData'
PlotExpression(object = pbmc_10Xassays, color.by = "scaled_mean_probs", dimred = "UMAP", color.scale = "viridis", legend.title = "Scaled mean prob.\n(min-max)") 

# Plot expression level of one or more features
## one 
PlotExpression(object = pbmc_10Xassays, color.by = "LGALS2", dimred = "UMAP")

## more than one
genes <- c("CD8A", "CCR7", "GZMA", "MS4A1", "CD27")
exp.plots <- lapply(X = genes, FUN = function(x) {
    PlotExpression(object = pbmc_10Xassays, color.by = x, dimred = "UMAP", scale.values = TRUE, point.size = 0.25, point.stroke = 0.25)
})
cowplot::plot_grid(plotlist = exp.plots, ncol = 5, align = "vh")

}
\keyword{Dimensional}
\keyword{reduction}
\keyword{visualization}
