% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{PrepareData}
\alias{PrepareData}
\alias{PrepareData.SingleCellExperiment}
\alias{PrepareData,SingleCellExperiment-method}
\title{Prepare \code{SingleCellExperiment} object for analysis}
\usage{
PrepareData.SingleCellExperiment(object)

\S4method{PrepareData}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class.}
}
\value{
An object of \code{SingleCellExperiment} class.
}
\description{
This function prepares the \code{SingleCellExperiment} object 
for analysis. The only required input is an object of class \code{SingleCellExperiment} 
with at least data in the \code{logcounts} slot.
}
\examples{
# Import package
library("SingleCellExperiment")

# Run function
pbmc_10Xassays <- PrepareData(pbmc_10Xassays)

}
\keyword{clean}
\keyword{data}
\keyword{normalized}
\keyword{prepare}
