% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceMappingMethods.R
\name{ReferenceMapping}
\alias{ReferenceMapping}
\alias{ReferenceMapping.SingleCellExperiment}
\alias{ReferenceMapping,SingleCellExperiment,SingleCellExperiment-method}
\title{Reference mapping}
\usage{
ReferenceMapping.SingleCellExperiment(ref, query, scale.query)

\S4method{ReferenceMapping}{SingleCellExperiment,SingleCellExperiment}(ref, query, scale.query = TRUE)
}
\arguments{
\item{ref}{An object of \code{SingleCellExperiment} class trained with ILoReg.}

\item{query}{An object of \code{SingleCellExperiment} class to project onto 
\code{ref}.}

\item{scale.query}{Should the query logcounts be scaled or not (logical). By 
default \code{TRUE}. Scale it if reference was scaled.}
}
\value{
An object of \code{SingleCellExperiment} class.
}
\description{
This functions allows to project new query data sets onto a reference built 
with ILoReg.
}
\keyword{ICP}
\keyword{LIBLINEAR}
\keyword{clustering}
\keyword{iterative}
\keyword{logistic}
\keyword{projection}
\keyword{regression}
