% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceMappingMethods.R
\name{ReferenceMapping}
\alias{ReferenceMapping}
\alias{ReferenceMapping.SingleCellExperiment}
\alias{ReferenceMapping,SingleCellExperiment,SingleCellExperiment-method}
\title{Reference mapping}
\usage{
ReferenceMapping.SingleCellExperiment(
  ref,
  query,
  ref.label,
  scale.query.by,
  project.umap,
  select.icp.models,
  k.nn,
  dimred.name.prefix
)

\S4method{ReferenceMapping}{SingleCellExperiment,SingleCellExperiment}(
  ref,
  query,
  ref.label,
  scale.query.by = NULL,
  project.umap = FALSE,
  select.icp.models = metadata(ref)$coralysis$pca.params$select.icp.tables,
  k.nn = 10,
  dimred.name.prefix = ""
)
}
\arguments{
\item{ref}{An object of \code{SingleCellExperiment} class trained with Coralysis
and after running \code{RunPCA(..., return.model = TRUE)} function.}

\item{query}{An object of \code{SingleCellExperiment} class to project onto 
\code{ref}.}

\item{ref.label}{A character cell metadata column name from the \code{ref} 
object to transfer to the queries.}

\item{scale.query.by}{Should the query data be scaled by \code{cell} or by 
\code{gene}. By default is \code{NULL}, i.e., is not scaled. Scale it if 
reference was scaled.}

\item{project.umap}{Project query data onto reference UMAP (logical). By 
default \code{FALSE}. If \code{TRUE}, the \code{ref} object needs to have a 
UMAP embedding obtained with \code{RunUMAP(..., return.model = TRUE)} function.}

\item{k.nn}{The number of \code{k} nearest neighbors to use in the classification
KNN algorithm used to transfer labels from the reference to queries (integer).
By default \code{10}.}

\item{dimred.name.prefix}{Dimensional reduction name prefix to add to the 
computed PCA and UMAP. By default nothing is added, i.e., 
\code{dimred.name.prefix = ""}.}

\item{selecte.icp.models}{Select the reference ICP models to use for query 
cluster probability prediction. By default \code{metadata(ref)$coralysis$pca.params$select.icp.tables}, 
i.e., the models selected to compute the reference PCA are selected. 
If \code{NULL} all are used. Otherwise a numeric vector should be given
to select the ICP models of interest.}
}
\value{
An object of \code{SingleCellExperiment} class.
}
\description{
This function allows to project new query data sets onto a reference built 
with Coralysis as well as transfer cell labels from the reference to queries.
}
\keyword{ICP}
\keyword{LIBLINEAR}
\keyword{clustering}
\keyword{iterative}
\keyword{logistic}
\keyword{projection}
\keyword{regression}
