% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RunPCA}
\alias{RunPCA}
\alias{RunPCA.SingleCellExperiment}
\alias{RunPCA,SingleCellExperiment-method}
\title{PCA transformation of the joint probability matrix}
\usage{
RunPCA.SingleCellExperiment(
  object,
  p,
  scale,
  center,
  threshold,
  method,
  return.model,
  select.icp.tables
)

\S4method{RunPCA}{SingleCellExperiment}(
  object,
  p = 50,
  scale = FALSE,
  center = TRUE,
  threshold = 0,
  method = "RSpectra",
  return.model = FALSE,
  select.icp.tables = NULL
)
}
\arguments{
\item{object}{object of \code{SingleCellExperiment} class}

\item{p}{a positive integer denoting the number of principal
components to calculate and select. Default is \code{50}.}

\item{scale}{a logical specifying whether the probabilities should be
standardized to unit-variance before running PCA. Default is \code{FALSE}.}

\item{center}{a logical specifying whether the probabilities should be
centered before running PCA. Default is \code{TRUE}.}

\item{threshold}{a thresfold for filtering out ICP runs before PCA with
the lower terminal projection accuracy below the threshold.
Default is \code{0}.}

\item{method}{A character specifying the PCA method. One of \code{"RSpectra"}
(default), \code{"irlba"} or \code{"stats"}.}

\item{return.model}{A logical specifying if the PCA model should or not be 
retrieved. By default \code{FALSE}. Only implemented for \code{method = "stats"}. 
If \code{TRUE}, the \code{method} is coerced to \code{"stats"}.}

\item{select.icp.tables}{Select the ICP cluster probability tables to perform 
PCA. By default \code{NULL}, i.e., all are used, except if the ICP tables were
obtained with the function \code{RunParallelDivisiveICP}, in which the ICP 
tables correspond to the last round of divisive clustering for every epoch.  
A vector of \code{integers} should be given otherwise.}
}
\value{
object of \code{SingleCellExperiment} class
}
\description{
Perform the PCA transformation of the joint probability matrix,
which reduces the dimensionality from k*L to p
}
\examples{
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays = list(logcounts = pbmc3k_500))
sce <- PrepareData(sce)
## These settings are just to accelerate the example, use the defaults.
sce <- RunParallelICP(sce,L=2,threads=1,C=0.1,k=5,r=1)
sce <- RunPCA(sce,p=5)


}
\keyword{PCA}
\keyword{eigendecomposition}
