% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RunPCA}
\alias{RunPCA}
\alias{RunPCA.SingleCellExperiment}
\alias{RunPCA,SingleCellExperiment-method}
\title{PCA transformation of the joint probability matrix}
\usage{
RunPCA.SingleCellExperiment(
  object,
  p,
  scale,
  center,
  threshold,
  method,
  return.model,
  select.icp.tables
)

\S4method{RunPCA}{SingleCellExperiment}(
  object,
  p = 50,
  scale = TRUE,
  center = TRUE,
  threshold = 0,
  method = "irlba",
  return.model = FALSE,
  select.icp.tables = NULL
)
}
\arguments{
\item{object}{object of \code{SingleCellExperiment} class}

\item{p}{a positive integer denoting the number of principal
components to calculate and select. Default is \code{50}.}

\item{scale}{a logical specifying whether the probabilities should be
standardized to unit-variance before running PCA. Default is \code{TRUE}.}

\item{center}{a logical specifying whether the probabilities should be
centered before running PCA. Default is \code{TRUE}.}

\item{threshold}{a thresfold for filtering out ICP runs before PCA with
the lower terminal projection accuracy below the threshold.
Default is \code{0}.}

\item{method}{A character specifying the PCA method. One of \code{"irlba"}
(default), \code{"RSpectra"} or \code{"stats"}.}

\item{return.model}{A logical specifying if the PCA model should or not be 
retrieved. By default \code{FALSE}. Only implemented for \code{method = "stats"}. 
If \code{TRUE}, the \code{method} is coerced to \code{"stats"}.}

\item{select.icp.tables}{Select the ICP cluster probability tables to perform 
PCA. By default \code{NULL}, i.e., all are used, except if the ICP tables were
obtained with the function \code{RunParallelDivisiveICP}, in which the ICP 
tables correspond to the last round of divisive clustering for every epoch.  
A vector of \code{integers} should be given otherwise.}
}
\value{
object of \code{SingleCellExperiment} class
}
\description{
Perform the PCA transformation of the joint probability matrix,
which reduces the dimensionality from k*L to p.
}
\examples{
# Import package
library("SingleCellExperiment")

# Prepare SCE object for analysis
pbmc_10Xassays <- PrepareData(pbmc_10Xassays)

# Multi-level integration - 'L = 4' just for highlighting purposes; use 'L=50' or greater
set.seed(123)
pbmc_10Xassays <- RunParallelDivisiveICP(object = pbmc_10Xassays, batch.label = "batch", L = 4) 

# Run PCA 
pbmc_10Xassays <- RunPCA(pbmc_10Xassays)

}
\keyword{PCA}
\keyword{eigendecomposition}
