% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RunTSNE}
\alias{RunTSNE}
\alias{RunTSNE.SingleCellExperiment}
\alias{RunTSNE,SingleCellExperiment-method}
\title{Barnes-Hut implementation of t-Distributed Stochastic Neighbor Embedding 
(t-SNE)}
\usage{
RunTSNE.SingleCellExperiment(
  object,
  dims,
  dimred.type,
  perplexity,
  dimred.name,
  ...
)

\S4method{RunTSNE}{SingleCellExperiment}(
  object,
  dims = NULL,
  dimred.type = "PCA",
  perplexity = 30,
  dimred.name = "TSNE",
  ...
)
}
\arguments{
\item{object}{Object of \code{SingleCellExperiment} class.}

\item{dims}{Dimensions to select from \code{dimred.type}. By default \code{NULL}, 
i.e., all the dimensions are selected. Provide a numeric vector to select a 
specific range, e.g., \code{dims = 1:10} to select the first 10 dimensions.}

\item{dimred.type}{Dimensional reduction type to use. By default \code{"PCA"}.}

\item{perplexity}{Perplexity of t-SNE.}

\item{dimred.name}{Dimensional reduction name given to the returned t-SNE. 
By default \code{"TSNE"}.}

\item{...}{Parameters to be passed to the \code{Rtsne} function. The parameters 
given should match the parameters accepted by the \code{Rtsne} function. Check 
possible parameters with \code{?Rtsne::Rtsne}.}
}
\value{
A \code{SingleCellExperiment} object.
}
\description{
Run nonlinear dimensionality reduction using t-SNE with the 
PCA-transformed consensus matrix as input.
}
\examples{
# Import package
library("SingleCellExperiment")

# Prepare SCE object for analysis
pbmc_10Xassays <- PrepareData(pbmc_10Xassays)

# Multi-level integration - 'L = 4' just for highlighting purposes; use 'L=50' or greater
set.seed(123)
pbmc_10Xassays <- RunParallelDivisiveICP(object = pbmc_10Xassays, batch.label = "batch", L = 4, threads = 1) 

# Run PCA 
set.seed(125) # to ensure reproducibility for the default 'irlba' method
pbmc_10Xassays <- RunPCA(pbmc_10Xassays, p = 10)

# Run t-SNE
set.seed(123)
pbmc_10Xassays <- RunTSNE(pbmc_10Xassays, dimred.type = "PCA")

}
\keyword{Barnes-Hut}
\keyword{Embedding}
\keyword{Neighbor}
\keyword{Stochastic}
\keyword{implementation}
\keyword{of}
\keyword{t-Distributed}
\keyword{t-SNE}
