% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RunTSNE}
\alias{RunTSNE}
\alias{RunTSNE.SingleCellExperiment}
\alias{RunTSNE,SingleCellExperiment-method}
\title{Barnes-Hut implementation of t-Distributed Stochastic
Neighbor Embedding (t-SNE)}
\usage{
RunTSNE.SingleCellExperiment(
  object,
  dims,
  dimred.type,
  perplexity,
  dimred.name,
  ...
)

\S4method{RunTSNE}{SingleCellExperiment}(
  object,
  dims = NULL,
  dimred.type = "PCA",
  perplexity = 30,
  dimred.name = "TSNE",
  ...
)
}
\arguments{
\item{object}{Object of \code{SingleCellExperiment} class.}

\item{dims}{Dimensions to select from \code{dimred.type}. By default \code{NULL}, 
i.e., all the dimensions are selected. Provide a numeric vector to select a 
specific range, e.g., \code{dims = 1:10} to select the first 10 dimensions.}

\item{dimred.type}{Dimensional reduction type to use. By default \code{"PCA"}.}

\item{perplexity}{Perplexity of t-SNE.}

\item{dimred.name}{Dimensional reduction name given to the returned t-SNE. 
By default \code{"TSNE"}.}

\item{...}{Parameters to be passed to the \code{Rtsne} function. The parameters 
given should match the parameters accepted by the \code{Rtsne} function. Check 
possible parameters with \code{?Rtsne::Rtsne}.}
}
\value{
object of \code{SingleCellExperiment} class
}
\description{
Run nonlinear dimensionality reduction using t-SNE with the
PCA-transformed consensus matrix as input.
}
\examples{
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays = list(logcounts = pbmc3k_500))
sce <- PrepareData(sce)
## These settings are just to accelerate the example, use the defaults.
sce <- RunParallelICP(sce,L=2,threads=1,C=0.1,k=5,r=1)
sce <- RunPCA(sce,p=5)
sce <- RunTSNE(sce)

}
\keyword{Barnes-Hut}
\keyword{Embedding}
\keyword{Neighbor}
\keyword{Stochastic}
\keyword{implementation}
\keyword{of}
\keyword{t-Distributed}
\keyword{t-SNE}
