% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RunUMAP}
\alias{RunUMAP}
\alias{RunUMAP.SingleCellExperiment}
\alias{RunUMAP,SingleCellExperiment-method}
\title{Uniform Manifold Approximation and Projection (UMAP)}
\usage{
RunUMAP.SingleCellExperiment(
  object,
  dims,
  dimred.type,
  return.model,
  umap.method,
  dimred.name,
  ...
)

\S4method{RunUMAP}{SingleCellExperiment}(
  object,
  dims = NULL,
  dimred.type = "PCA",
  return.model = FALSE,
  umap.method = "umap",
  dimred.name = "UMAP",
  ...
)
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class.}

\item{dims}{Dimensions to select from \code{dimred.type}. By default \code{NULL}, 
i.e., all the dimensions are selected. Provide a numeric vector to select a 
specific range, e.g., \code{dims = 1:10} to select the first 10 dimensions.}

\item{dimred.type}{Dimensional reduction type to use. By default \code{"PCA"}.}

\item{return.model}{Return UMAP model. By default \code{FALSE}.}

\item{umap.method}{UMAP method to use: \code{"umap"} or \code{"uwot"}. 
By default \code{"umap"}.}

\item{dimred.name}{Dimensional reduction name given to the returned UMAP. 
By default \code{"UMAP"}.}

\item{...}{Parameters to be passed to the \code{umap} function. The parameters 
given should match the parameters accepted by the \code{umap} function depending
on the \code{umap.method} given. Check possible parameters with \code{?umap::umap} 
or \code{?uwot::umap} depending if \code{umap.method} is \code{"umap"} or 
\code{"uwot"}.}
}
\value{
A \code{SingleCellExperiment} object.
}
\description{
Run nonlinear dimensionality reduction using UMAP with a dimensional 
reduction as input.
}
\examples{
# Import package
library("SingleCellExperiment")

# Prepare SCE object for analysis
pbmc_10Xassays <- PrepareData(pbmc_10Xassays)

# Multi-level integration - 'L = 4' just for highlighting purposes; use 'L=50' or greater
set.seed(123)
pbmc_10Xassays <- RunParallelDivisiveICP(object = pbmc_10Xassays, batch.label = "batch", L = 4, threads = 1) 

# Run PCA 
set.seed(125) # to ensure reproducibility for the default 'irlba' method
pbmc_10Xassays <- RunPCA(pbmc_10Xassays, p = 10)

# Run UMAP
set.seed(123)
pbmc_10Xassays <- RunUMAP(pbmc_10Xassays, dimred.type = "PCA")

}
\keyword{Approximation}
\keyword{Manifold}
\keyword{Projection}
\keyword{UMAP}
\keyword{Uniform}
\keyword{and}
