% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RunUMAP}
\alias{RunUMAP}
\alias{RunUMAP.SingleCellExperiment}
\alias{RunUMAP,SingleCellExperiment-method}
\title{Uniform Manifold Approximation and Projection (UMAP)}
\usage{
RunUMAP.SingleCellExperiment(
  object,
  dimred.type,
  return.model,
  umap.method,
  dimred.name,
  ...
)

\S4method{RunUMAP}{SingleCellExperiment}(
  object,
  dimred.type = "PCA",
  return.model = FALSE,
  umap.method = "umap",
  dimred.name = "UMAP",
  ...
)
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class}

\item{dimred.type}{Dimensional reduction type to use. By default \code{"PCA"}.}

\item{return.model}{Return UMAP model. By default \code{FALSE}.}

\item{umap.method}{UMAP method to use: \code{"umap"} or \code{"uwot"}. 
By default \code{"umap"}.}

\item{dimred.name}{Dimensional reduction name given to the returned UMAP. 
By default \code{"UMAP"}.}

\item{...}{Parameters to be passed to the \code{umap} function. The parameters 
given should match the parameters accepted by the \code{umap} function depending
on the \code{umap.method} given. Check possible parameters with \code{?umap::umap} 
or \code{?uwot::umap} depending if \code{umap.method} is \code{"umap"} or 
\code{"uwot"}.}
}
\value{
Object of \code{SingleCellExperiment} class.
}
\description{
Run nonlinear dimensionality reduction using UMAP with a dimensional reduction as input.
}
\examples{
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays = list(logcounts = pbmc3k_500))
sce <- PrepareILoReg2(sce)
## These settings are just to accelerate the example, use the defaults.
sce <- RunParallelICP(sce,L=2,threads=1,C=0.1,k=5,r=1)
sce <- RunPCA(sce,p=5)
sce <- RunUMAP(sce)

}
\keyword{Approximation}
\keyword{Manifold}
\keyword{Projection}
\keyword{UMAP}
\keyword{Uniform}
\keyword{and}
