% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{SummariseCellClusterProbability}
\alias{SummariseCellClusterProbability}
\alias{SummariseCellClusterProbability.SingleCellExperiment}
\alias{SummariseCellClusterProbability,SingleCellExperiment-method}
\title{Summarise ICP cell cluster probability}
\usage{
SummariseCellClusterProbability.SingleCellExperiment(
  object,
  icp.run,
  icp.round,
  funs,
  scale.funs,
  save.in.sce
)

\S4method{SummariseCellClusterProbability}{SingleCellExperiment}(
  object,
  icp.run = NULL,
  icp.round = NULL,
  funs = c("mean", "median"),
  scale.funs = TRUE,
  save.in.sce = TRUE
)
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class with ICP cell 
cluster probability tables saved in \code{metadata(object)$iloreg$joint.probability}. 
After running one of \code{RunParallelICP} or \code{RunParallelDivisiveICP}.}

\item{icp.run}{ICP run(s) to retrieve from \code{metadata(object)$iloreg$joint.probability}. 
By default \code{NULL}, i.e., all are retrieved. Specify a numeric vector to 
retrieve a specific set of tables.}

\item{icp.round}{ICP round(s) to retrieve from \code{metadata(object)$iloreg$joint.probability}. 
By default \code{NULL}, i.e., all are retrieved. Only relevant if probabilities
were obtained with the function \code{RunParallelDivisiveICP}, i.e., divisive ICP
was performed. Otherwise it is ignored and internally assumed as \code{icp.round = 1}, 
i.e., only one round.}

\item{funs}{Functions to summarise ICP cell cluster probability: \code{"mean"} 
and/or \code{"median"}. By default \code{c("mean", "median")}, i.e, both mean
and median are calculated. Set to \code{NULL} to not estimate any.}

\item{scale.funs}{Scale in the range 0-1 the summarised probability obtained with 
\code{funs}. By default \code{TRUE}, i.e., summarised probability will be scaled
in the 0-1 range.}

\item{save.in.sce}{Save the data frame into the cell metadata from the 
\code{SingleCellExperiment} object or return the data frame. By default \code{TRUE}, 
i.e., the summary of probabilities retrieved is save in the SCE object in 
\code{colData(object)}.}
}
\value{
A data frame or a SingleCellExperiment object with ICP cell cluster probability summarised.
}
\description{
Summarise ICP cell cluster probability table(s)
}
\keyword{Summarise}
\keyword{cell}
\keyword{cluster}
\keyword{probability}
