% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellStateMethods.R
\name{TabulateCellBinsByGroup}
\alias{TabulateCellBinsByGroup}
\alias{TabulateCellBinsByGroup.SingleCellExperiment}
\alias{TabulateCellBinsByGroup,SingleCellExperiment-method}
\title{Tabulate cell bins by group}
\usage{
TabulateCellBinsByGroup.SingleCellExperiment(object, group, relative, margin)

\S4method{TabulateCellBinsByGroup}{SingleCellExperiment}(object, group, relative = FALSE, margin = 1)
}
\arguments{
\item{object}{An object of \code{SingleCellExperiment} class obtained with the 
function \code{BinCellClusterProbability()}.}

\item{group}{Character specifying the \code{colData} variable from the \code{SingleCellExperiment} 
object provided to the function \code{BinCellClusterProbability()} to use as 
categorical group variable.}

\item{relative}{Logical specifying if relative proportions of cell bins per 
\code{group} should be returned. By default \code{FALSE}, i.e., absolute values
are returned.}

\item{margin}{If \code{relative} is \code{TRUE}, proportions should be calculated 
by: rows (\code{1}, the default); columns (\code{2}); or overall (\code{NULL}).}
}
\value{
A list of tables with the frequency of cells per bin of cell cluster probability by group for each label.
}
\description{
Frequency of cells per cell cluster probability bin by group for each label. 
The label has to be specified beforehand to the function \code{BinCellClusterProbability()}.
}
\keyword{Table}
\keyword{bins}
\keyword{cell}
\keyword{group}
