% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizMethods.R
\name{VlnPlot}
\alias{VlnPlot}
\alias{VlnPlot.SingleCellExperiment}
\alias{VlnPlot,SingleCellExperiment-method}
\title{Visualization of feature expression using violin plots}
\usage{
VlnPlot.SingleCellExperiment(
  object,
  clustering.label,
  features,
  return.plot,
  rotate.x.axis.labels
)

\S4method{VlnPlot}{SingleCellExperiment}(
  object,
  clustering.label,
  features,
  return.plot = FALSE,
  rotate.x.axis.labels = FALSE
)
}
\arguments{
\item{object}{of \code{SingleCellExperiment} class}

\item{clustering.label}{A variable name (of class \code{character}) available 
in the cell metadata \code{colData(object)} with the clustering labels 
(\code{character} or \code{factor}) to use.}

\item{features}{Feature names to plot by cluster (\code{character}) matching 
\code{row.names(object)}.}

\item{return.plot}{return.plot whether to return the \code{ggplot2} object. 
Default is \code{FALSE}.}

\item{rotate.x.axis.labels}{a logical denoting whether the x-axis labels should 
be rotated 90 degrees or just draw it. Default is \code{FALSE}.}
}
\value{
A \code{ggplot2} object if \code{return.plot=TRUE}.
}
\description{
The \code{VlnPlot} function enables visualizing expression levels 
of feature(s), across clusters using violin plots.
}
\examples{
# Import package
library("SingleCellExperiment")

# Prepare SCE object for analysis
pbmc_10Xassays <- PrepareData(pbmc_10Xassays)

# Multi-level integration - 'L = 4' just for highlighting purposes; use 'L=50' or greater
set.seed(123)
pbmc_10Xassays <- RunParallelDivisiveICP(object = pbmc_10Xassays, batch.label = "batch", L = 4, threads = 1) 

# Run PCA 
set.seed(125) # to ensure reproducibility for the default 'irlba' method
pbmc_10Xassays <- RunPCA(pbmc_10Xassays, p = 10)

# Plot features by cell type (give a clustering instead)
genes <- c("CD8A", "CCR7", "GZMA", "MS4A1", "CD27")
VlnPlot(pbmc_10Xassays, clustering.label = "cell_type", features = genes, rotate.x.axis.labels = TRUE)

}
\keyword{plot}
\keyword{violin}
